note: branched off https://github.com/pelias/interpolation/pull/157, please merge that PR first.
diff: https://github.com/pelias/interpolation/compare/node_9...performance

This PR is the result of some work to improve the performance of the `/search/geojson` endpoint which is heavily used by Pelias.

My initial testing confirmed that the performance was very poor, I wasn't able to test any concurrency because it was returning responses in `~681.705 ms` at a concurrency of 1!

With a little query optimization and the addition of another field to a covering index brings the performance back in the range of acceptability for production use:

### latency

At a concurrency of 250/s I was able to achieve the following latency distribution on my development laptop:

```
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    5.22ms
 75.000%    6.02ms
 90.000%    6.82ms
 99.000%   17.81ms
 99.900%   34.11ms
 99.990%   37.98ms
 99.999%   37.98ms
100.000%   37.98ms
```

It is capable of doing more QPS, when I tested it at 350/s it has a pretty similar distribution (+5ms) but when trying to push it above 450/s there is congestion and the resulting latency graph contains a bunch of slow queries and a far worse average latency.

### index updates

The `names_name_idx` index in the `streets.db` is actually causing performance to greatly suffer, so deleting it will result in a significant performance improvement:

```
sqlite3 street.db <<< 'DROP INDEX IF EXISTS names_name_idx;'
```

... but since it might be required by other queries, I would strongly recommend recreating it with an additional field added:

```
sqlite3 street.db <<< 'CREATE INDEX IF NOT EXISTS names_name_idx ON names(name, id);'
```

All new databases created after this PR is merged will automatically have the correct index, databases created prior to this change will need to be patched using the two commands above.

Changes to the index may result in fragmentation within the database file, this will likely have no effect on performance but if you would like to reclaim some disk space and don't mind waiting for the database to be re-written un-fragmented, you can:

```
export SQLITE_TMPDIR=/disk/with/enough/free/space
sqlite3 street.db <<< 'VACUUM;'
```

Note: this is not required and will consume temporary disk space equal to the size of the database file (~8GB) and will take some time to execute, you will probably reclaim about 100MB by doing this.