@missinglink 2018-06-18
I tested this today on a `c5d.2xlarge` and it was capable of handling up to 1400 requests per-sec:

```
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.30ms
 75.000%    2.83ms
 90.000%    3.38ms
 99.000%    7.47ms
 99.900%   10.91ms
 99.990%   13.53ms
 99.999%   14.34ms
100.000%   14.34ms

...
  13989 requests in 10.00s, 5.95MB read
Requests/sec:   1398.75
Transfer/sec:    609.22KB
```

I'm guessing the exceptional performance here is a combination of the high clock speed of those machines, coupled with the fast local drives and the fact that I'm just hitting the same URL over-and-over which is being cached by the filesystem and so no disk reads are happening.

Either way, it's great news as one server can handle a lot of load now :)

Also noticed that 1400/s is basically the limit, once I push it to 1500/s it very quickly falls off the cliff:

```
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%  499.20ms
 75.000%  599.04ms
 90.000%  737.79ms
 99.000%  856.06ms
 99.900%  880.64ms
 99.990%  884.73ms
 99.999%  885.25ms
100.000%  885.25ms

  13723 requests in 10.00s, 5.84MB read
Requests/sec:   1372.20
Transfer/sec:    597.66KB
```