@missinglink 2018-06-18
I have patched the `street.db` & `street.db.gz` files in s3 buckets for both `nextzen` & `geocodeearth`, I also kept a copy of the database prior to the index changes just in case.

```
2018-06-18 10:19:05        159 about.pre.perf
2018-04-15 22:57:53 45494140928 address.db
2018-04-15 22:58:51 18125199126 address.db.gz
2018-06-18 10:17:52 6739410944 street.db
2018-06-18 10:18:34 3173994096 street.db.gz
2018-06-18 10:17:17 3257081503 street.db.gz.pre.perf
2018-06-18 10:16:33 6400380928 street.db.pre.perf
```