@Binternet 2018-06-18
haven't tested it (yet) but If it passed your tests it's fine by me =)